test: use specified paths for testing generators
The createGenerator() provided by yeoman-generator test/helpers
doesn't allow for specifying the path of a generator, so it requires
that the generator be in a generator-* parent directory. This is
rarely true with Jenkins matrix jobs, so the tests always fail.

This replacement performs all of the operations used by the tests
in generator-loopback while allowing the path to be specified.