Add prompt for default values on model properties
This commit adds a prompt for a default value when a new property is created on a model.  Currently, it supports the boolean values of true and false, the special cases of "guid", "uuid", "now" on date/datetime, as well as strings.  

I'm not super in love with this nested if/else structure, but since everything comes back from the command line as a string which we have to parse (based on the property's type), I'm not immediately seeing a better way to do this.  Additionally "guid", "uuid", and "now" are special cases as they set the property "defaultFn" as opposed to just "default".  It's not elegant, but it does work.